Address feedback from PR review
